add command line in setup.py
