+++ ml/cvt_smallbox/command_line.py
@@ -1,5 +1,3 @@
-
-
 def main():
-    from . import cvt_smallbox
-    cvt_smallbox.process()
+    from .code import process
+    process()
